Added TypeScript configuration. Added Webpack configuration. Move source code to /src folder.
