+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "sweet-data",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "author": "Brian Gerhards (https://github.com/bgerhards)",
   "description": "plugin to pretty-print or minify XML, JSON, CSS and SQL files",
   "contributors": [
     "Brian Gerhards (https://github.com/bgerhards)",
-    "Alex Staroselsky (https://github.com/alstar01)"
+    "Alex Staroselsky (https://github.com/alexstaroselsky)"
   ],
   "repository": {
     "type": "git",
@@ -29,8 +29,12 @@
     "node": ">=0.4.9"
   },
   "devDependencies": {
+    "awesome-typescript-loader": "^3.4.1",
     "mocha": "^3.4.2",
-    "should": "^11.2.1"
+    "should": "^11.2.1",
+    "shx": "^0.2.2",
+    "typescript": "^2.6.2",
+    "webpack": "^3.10.0"
   },
   "files": [
     "README.md",
@@ -38,6 +42,8 @@
     "lib/"
   ],
   "scripts": {
+    "clean": "shx rm -rf _bundles lib lib-esm",
+    "build": "npm run clean && tsc && tsc -m es6 --outDir lib-esm && webpack",
     "test": "mocha",
     "test-watch": "mocha -w"
   },
