+++ src/index.ts
@@ -2,11 +2,11 @@ function pp(steps, stepChar) {
 	steps = typeof steps !== 'undefined' ? steps : 2;
 	stepChar = typeof stepChar !== 'undefined' ? stepChar : ' ';
 	this.setStep(steps, stepChar);
-};	
+};
 
 pp.prototype.setStep = function(steps, stepChar){
 	this.shift = ['\n']; // array of shifts
-	
+
 	switch(stepChar){
 		case 'SPACE': this.stepChar = ' '; break;
 		case 'TAB': this.stepChar = '	'; break;
@@ -21,17 +21,17 @@ pp.prototype.setStep = function(steps, stepChar){
 			this.step += this.stepChar;
 		}
 	}
-	
+
 	var maxdeep = 100, // nesting level
       ix = 0;
 
 	// initialize array with shifts //
 	for(ix=0;ix<maxdeep;ix++){
-		this.shift.push(this.shift[ix]+this.step); 
+		this.shift.push(this.shift[ix]+this.step);
 	}
 
 }
-	
+
 
 // ----------------------- beautify section -----------------------------------------------
 pp.prototype.xml = require('./xml').xml;
